post로 사용자추천 출력 recommand 새로작성 mapper.xml포함

+++ src/main/java/com/team/winey/recommand/RecommandMapper.java
@@ -8,8 +8,7 @@ import java.util.List;
 
 @Mapper
 public interface RecommandMapper {
-
-    List<Integer> selRecommand(RecommandRes res);
+    List<Long> selRecommand(RecommandRes res);
     int insUserinfo(UserinfoDto dto);
     List<Integer> selUserinfo(SelRecommandDto dto);
 }
